♻️ [Refactoring]: #221 Copilot レビュー指摘 4 件を修正
- resolve_enabled のパラメータ名を plugin_name → install_id にリネーム
  （実際に渡すのは dir_name/cache_key であり旧名は誤解を招く）
- WireSource::Github → WireSource::GitHub に統一（application::Source::GitHub との表記一致。rename_all="lowercase" により serialize 結果は不変）
- Wire / WireAuthor / WireSource / WireComponents を commands/info/wire.rs に抽出
  （json.rs と yaml.rs 間の意図しないドリフトを防止）
- build_plugin_info が InstalledPlugin に渡す marketplace を Option<String> のまま保つ
  （list_installed_plugins と同一セマンティクス、GitHub は None 表現を維持）

Refs #221
